Exclude _ScCbts cookie from managed firewall rules (#153)
The `_ScCbts` cookie is not used by this site, however other parts of the DfE estate set it at an excessively wide level. This is used for Snapchat conversion tracking in Get into Teaching, for example. The cookie, when set, has an expiry of 6 days, and contains a JSON object. This is often enough to trigger the managed ruleset as suspicious.

This change excludes that cookie from weight considerations as part of the managed ruleset, meaning a suspect body will not contribute to a reject decision. While it would be better to prevent this cookie being set in the first place, the behaviour of the snapchat SDK is to choose the domain to use based on the widest that will work. As education.gov.uk is not in the public suffix list, this is the one that succeeds. Services on service.gov.uk do not have this problem, as that domain is on the suffix list.